requirements.txt
python-docx
memory_profiler 